Make always re-generating the test_app opt-in
Generating the app each time the specs are ran is very time expensive
and should be avoided if possible, for those situations in which this
is necessary instructions have been added to the Readme and Changelog.

For further info on how the `directory` rule works in Rake see:
https://ruby.github.io/rake/Rake/DSL.html#method-i-directory

Partially reverts deb9243c, 93ce7712, and cba90008.