Fix eval_gemfile call breaking Dependabot analysis
Dependabot tries to read files that are passed as arguments to
eval_gemfile[1], but in our case the file doesn't actually exist.
This causes Dependabot to throw an error and not analyze the
project.

Since the logic behind eval_gemfile analysis just uses a simple
parser[2], using `send` instead of calling the method directly is
enough to prevent Dependabot from recognizing the method call.

We have also opened an issue on the Dependabot repository[3], but
there are a lot of edge cases to account for and a lot of different
approaches, so for the time being this quick fix will be enough.

[1]: https://github.com/dependabot/dependabot-core/blob/8831793d71c6a6a17eb21168a9015c8c526173b1/lib/dependabot/file_fetchers/ruby/bundler/child_gemfile_finder.rb#L31
[2]: https://github.com/dependabot/dependabot-core/blob/8831793d71c6a6a17eb21168a9015c8c526173b1/lib/dependabot/file_fetchers/ruby/bundler/child_gemfile_finder.rb#L51
[3]: https://github.com/dependabot/dependabot-core/issues/1658