correct Gemfile for sandbox generation
To run the sandbox correctly we need to add Solidus to the Gemfile as well. This commit adds the option to run the sandbox with a specific Solidus `SOLIDUS_BRANCH`, by default it will be using the `master` branch.

The order of the `Gemfile` also matters, make sure we have `solidus` as the first gem and the extension under development as the last entry for the sandbox.