Fix running the rails generator inside the extension
Since extensions are engines, my expectation would be that the bin/rails command
is working as well. This way I can generate my migrations, models etc.
This is a Rails convention as well.

Also see https://guides.rubyonrails.org/engines.html#bin-directory

There are some problems with having a `bin/rails` file that loads the `ENGINE_PATH`
and `ENGINE_ROOT`, so we have to rename the executable.

This commit introduces the `bin/r` executable to run as the `bin/rails` command that
will add the option to generate models, migrations etc. inside the extension.

It also introduces the `bin/sandbox_rails` executable to be able to run the rails
commands on the generated sandbox rails application.