Remove double auth installation in sandbox
In Solidus master, and starting from 2.11, we'll have the solidus
install script to install auth if auto-accept=true.

The option added by this commit allows to skip the installation
of solidus_auth_devise in spree:install, since we are already
doing this.

We could also remove the manual solidus_auth_devise installation
but it would be not possibile to generate a sandbox for past
versions of Solidus. We can switch to auto install solidus when
we'll drop support for 2.10 though.