Stop enabling all new RuboCop cops
We want all Solidus developers to have a consistent experience when
linting their extensions. The easiest way to do this, _without_ forcing
a RuboCop version on them is to stop automatically enabling new RuboCop
cops.

Instead, we'll only use the cops that are in the current major, and
wait for the next one to enable new cops.