Do not raise if source_code_uri is missing in the gemspec
That metadata is useful to automatically generate the CHANGELOG
based on GitHub activity, using Oktokit and github_changelog_generator.

But a lot of extensions do not have that metadata set. We can be more
kind by not making specs ran with rake (which are unrelated) fail when
that information is missing.