Always load solidus core factories when testing extensions
We need this because most of the extensions have only

require 'extension_name/factories'

in their spec_helper. We need to manually load solidus factories
for extensions that didn't switched to load_for() yet.

But now, core emits a deprecation warning when trying to load
factories using require, so we can aviod to print that deprecation
warning and use a custom message that explain how to migrate
extensions in this transition period.

When we have all extensions updated with load_for() for factories,
we can safely remove the manual require of solidus core factories
and the cutom message message along with it.