Support solidus_auth_devise in the sandbox script
Don't add solidus_auth_devise to the sandbox app if it's the current
extension.

In 3017ed64 solidus_auth_devise was added directly to the gemfile in
order to support solidus prior to v2.11. Since we don't need to
support it anymore we can safely use the official
`--with-authentication` option now.