Streamline CircleCI configuration
The previous configuration had a special way of updating bundler that
was necessary 21 months ago, but probably is not anymore. This changes
the configuration for a bog-standard configuration that should also run
our specs.

The previous configuration also failed at cleaning up, with the
following error:

```
https://github.com/solidusio/solidus (at v4.3@3ded512) is not yet checked out. Run `bundle install` first.

Exited with code exit status 11
```

https://app.circleci.com/pipelines/github/solidusio/solidus_dev_support/913/workflows/d9281ce7-fe62-45d8-b309-4f287b890924/jobs/2953