Allow passing a user class to the test app generator
Currently, solidus_auth_devise is tested with a dummy app that specifies
`Spree::LegacyUser` in its `spree.rb initializer. That's not great, but
there's no way of fixing it if the rake task install does not pass
through the user class option.

The way to use this feature would be in an extension's Rakefile:

```rb
SolidusDevSupport::RakeTasks.install(user_class: "Spree::User")
```