Generate Manifest file early in sandbox generation
The sandbox wants to install Solidus and migrate before running the
install generator. That's bad, because migrating requires the app to
boot, and booting fails because Sprockets fails with an error if the
`manifest.js` file is not present.

Since the file is quite simple, we can just pass it into a heredoc. This
is needed for Rails 8, where the install generator won't create the file
for us.